backend/back2/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
pandas
scikit-learn
mlflow
scipy
python-multipart
aiofiles
httpx
fastapi-cors
